fixing

+++ commands/timer.js
@@ -4,6 +4,9 @@
 
 module.exports = {
   run: async (client, message, [ time ]) => {
+
+    if (!message.member.hasPermission('MANAGE_MESSAGES')) return message.reply('você não tem permissão para usar esse comando!')
+
     await message.delete()
     const messages = await message.channel.fetchMessages({ limit: 100 })
     const userMessages = messages.filter((m) => m.author === message.author && m.deletable)
